% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omg-package.R
\docType{package}
\name{omg-package}
\alias{omg-package}
\title{omg: Ortholog Marker Gene cross-species cell type annotation}
\description{
The main entry point is \code{\link{omg}}, which reproduces the
\code{Omg_annotation.R} command-line pipeline as a single R call.
Helper functions \code{\link{omg_list_species}} and
\code{\link{omg_check_reference}} support adding new species.
}
\author{
\strong{Maintainer}: Tran Chau \email{chauducanh@gmail.com}

}
\keyword{internal}
